refactor: move fonts and colors to index
