--- src/constants/styles/colors.ts
@@ -1,9 +1,0 @@
-export const colors = {
-  primaryGreen: '#224229',
-  lightGreen: '#F4F6F5',
-  mediumGreen: '#59695C',
-  darkGreen: '#162C1B',
-  greyshGreen: '#DAE4DC',
-  white: '#FFFFFF',
-  accent: '#F89F34',
-};
